clean up HostDB interface
