+++ modules/host/contractcreation.go
@@ -2,17 +2,17 @@ package host
 
 import (
 	"errors"
-	"fmt"
-	"io"
+	// "fmt"
+	// "io"
 	"net"
-	"os"
-	"path/filepath"
+	// "os"
+	// "path/filepath"
 	"strconv"
 
 	"github.com/NebulousLabs/Sia/consensus"
-	"github.com/NebulousLabs/Sia/encoding"
-	"github.com/NebulousLabs/Sia/hash"
-	"github.com/NebulousLabs/Sia/modules"
+	// "github.com/NebulousLabs/Sia/encoding"
+	// "github.com/NebulousLabs/Sia/hash"
+	// "github.com/NebulousLabs/Sia/modules"
 )
 
 var (
@@ -43,83 +43,85 @@ func (h *Host) nextFilename() string {
 //
 // TODO: Make the host able to parse multiple contracts at once.
 func (h *Host) considerContract(t consensus.Transaction, startBlock consensus.BlockHeight) (updatedTransaction consensus.Transaction, err error) {
-	// Check that there is exactly one file contract.
-	if len(t.FileContracts) != 1 {
-		err = errors.New("transaction must have exactly one contract")
-		return
-	}
-	if startBlock > h.state.Height()+20 {
-		err = errors.New("startBlock is too far in the future")
-		return
-	}
+	/*
+		// Check that there is exactly one file contract.
+		if len(t.FileContracts) != 1 {
+			err = errors.New("transaction must have exactly one contract")
+			return
+		}
+		if startBlock > h.state.Height()+20 {
+			err = errors.New("startBlock is too far in the future")
+			return
+		}
 
-	// These variables are here for convenience.
-	contract := t.FileContracts[0]
-	window := contract.End - contract.Start
-	duration := contract.End - startBlock
-	fileSize := contract.FileSize
+		// These variables are here for convenience.
+		contract := t.FileContracts[0]
+		window := contract.End - contract.Start
+		duration := contract.End - startBlock
+		fileSize := contract.FileSize
 
-	// Check that the file size listed in the contract is in bounds.
-	if fileSize < h.announcement.MinFilesize || fileSize > h.announcement.MaxFilesize {
-		err = fmt.Errorf("file is of incorrect size - filesize %v, min %v, max %v", fileSize, h.announcement.MinFilesize, h.announcement.MaxFilesize)
-		return
-	}
-	// Check that there is space for the file.
-	if fileSize > uint64(h.spaceRemaining) {
-		err = HostCapacityErr
-		return
-	}
-	// Check that the duration of the contract is in bounds.
-	if duration < h.announcement.MinDuration || duration > h.announcement.MaxDuration {
-		err = errors.New("contract duration is out of bounds")
-		return
-	}
-	// Check that the window is large enough.
-	if window < h.announcement.MinWindow {
-		err = errors.New("challenge window is not large enough")
-		return
-	}
-	// Outputs for successful proofs need to go to the correct address.
-	if contract.ValidProofAddress != h.announcement.CoinAddress {
-		err = errors.New("coins are not paying out to correct address")
-		return
-	}
-	// Output for failed proofs needs to be the 0 address.
-	emptyAddress := consensus.CoinAddress{}
-	if contract.MissedProofAddress != emptyAddress {
-		err = errors.New("burn payout needs to go to the empty address")
-		return
-	}
+		// Check that the file size listed in the contract is in bounds.
+		if fileSize < h.announcement.MinFilesize || fileSize > h.announcement.MaxFilesize {
+			err = fmt.Errorf("file is of incorrect size - filesize %v, min %v, max %v", fileSize, h.announcement.MinFilesize, h.announcement.MaxFilesize)
+			return
+		}
+		// Check that there is space for the file.
+		if fileSize > uint64(h.spaceRemaining) {
+			err = HostCapacityErr
+			return
+		}
+		// Check that the duration of the contract is in bounds.
+		if duration < h.announcement.MinDuration || duration > h.announcement.MaxDuration {
+			err = errors.New("contract duration is out of bounds")
+			return
+		}
+		// Check that the window is large enough.
+		if window < h.announcement.MinWindow {
+			err = errors.New("challenge window is not large enough")
+			return
+		}
+		// Outputs for successful proofs need to go to the correct address.
+		if contract.ValidProofAddress != h.announcement.CoinAddress {
+			err = errors.New("coins are not paying out to correct address")
+			return
+		}
+		// Output for failed proofs needs to be the 0 address.
+		emptyAddress := consensus.CoinAddress{}
+		if contract.MissedProofAddress != emptyAddress {
+			err = errors.New("burn payout needs to go to the empty address")
+			return
+		}
 
-	// Verify that the contract fund covers the payout and burn for the whole
-	// duration.
-	requiredFund := (h.announcement.Price + h.announcement.Burn) * consensus.Currency(duration) * consensus.Currency(fileSize)
-	if contract.Payout != requiredFund {
-		err = errors.New("ContractFund does not match the terms of service.")
-		return
-	}
+		// Verify that the contract fund covers the payout and burn for the whole
+		// duration.
+		requiredFund := (h.announcement.Price + h.announcement.Burn) * consensus.Currency(duration) * consensus.Currency(fileSize)
+		if contract.Payout != requiredFund {
+			err = errors.New("ContractFund does not match the terms of service.")
+			return
+		}
 
-	// Add enough funds to the transaction to cover the penalty half of the
-	// agreement.
-	penalty := h.announcement.Burn * consensus.Currency(fileSize) * consensus.Currency(duration)
-	id, err := h.wallet.RegisterTransaction(t)
-	if err != nil {
-		err = HostCapacityErr // hide the fact that the host is having wallet issues.
-		return
-	}
-	err = h.wallet.FundTransaction(id, penalty)
-	if err != nil {
-		err = HostCapacityErr // hide the fact that the host is having wallet issues.
-		return
-	}
-	updatedTransaction, err = h.wallet.SignTransaction(id, true)
-	if err != nil {
-		err = HostCapacityErr // hide the fact that the host is having wallet issues.
-		return
-	}
+		// Add enough funds to the transaction to cover the penalty half of the
+		// agreement.
+		penalty := h.announcement.Burn * consensus.Currency(fileSize) * consensus.Currency(duration)
+		id, err := h.wallet.RegisterTransaction(t)
+		if err != nil {
+			err = HostCapacityErr // hide the fact that the host is having wallet issues.
+			return
+		}
+		err = h.wallet.FundTransaction(id, penalty)
+		if err != nil {
+			err = HostCapacityErr // hide the fact that the host is having wallet issues.
+			return
+		}
+		updatedTransaction, err = h.wallet.SignTransaction(id, true)
+		if err != nil {
+			err = HostCapacityErr // hide the fact that the host is having wallet issues.
+			return
+		}
 
-	// Update the amount of space the host has for sale.
-	h.spaceRemaining -= int64(fileSize)
+		// Update the amount of space the host has for sale.
+		h.spaceRemaining -= int64(fileSize)
+	*/
 
 	return
 }
@@ -164,92 +166,94 @@ func (h *Host) considerContract(t consensus.Transaction, startBlock consensus.Bl
 // TODO: This functions error handling isn't safe, reveals too much info to the
 // other party.
 func (h *Host) NegotiateContract(conn net.Conn) (err error) {
-	// Read the transaction from the connection.
-	var txn consensus.Transaction
-	err = encoding.ReadObject(conn, &txn, maxContractLen)
-	if err != nil {
-		return
-	}
-	var startBlock consensus.BlockHeight
-	err = encoding.ReadObject(conn, &startBlock, maxContractLen) // what should be the maxlen here?
-	if err != nil {
-		return
-	}
-	contract := txn.FileContracts[0]
+	/*
+		// Read the transaction from the connection.
+		var txn consensus.Transaction
+		err = encoding.ReadObject(conn, &txn, maxContractLen)
+		if err != nil {
+			return
+		}
+		var startBlock consensus.BlockHeight
+		err = encoding.ReadObject(conn, &startBlock, maxContractLen) // what should be the maxlen here?
+		if err != nil {
+			return
+		}
+		contract := txn.FileContracts[0]
 
-	// Check that the contained FileContract fits host criteria for taking
-	// files, replying with the error if there's a problem.
-	h.mu.Lock()
-	txn, err = h.considerContract(txn, startBlock)
-	h.mu.Unlock()
-	if err != nil {
-		_, err = encoding.WriteObject(conn, err.Error())
-		return
-	}
-	_, err = encoding.WriteObject(conn, modules.AcceptContractResponse)
-	if err != nil {
-		return
-	}
-
-	// Create file.
-	h.mu.Lock()
-	filename := h.nextFilename()
-	fullname := filepath.Join(h.hostDir, filename)
-	h.mu.Unlock()
-	file, err := os.Create(fullname)
-	if err != nil {
-		return
-	}
-	defer file.Close()
-
-	// If there's an error upon return, delete the file that's been created.
-	defer func() {
+		// Check that the contained FileContract fits host criteria for taking
+		// files, replying with the error if there's a problem.
+		h.mu.Lock()
+		txn, err = h.considerContract(txn, startBlock)
+		h.mu.Unlock()
 		if err != nil {
 			_, err = encoding.WriteObject(conn, err.Error())
-			os.Remove(fullname)
-			h.mu.Lock()
-			h.spaceRemaining -= int64(contract.FileSize)
-			h.mu.Unlock()
+			return
 		}
-	}()
+		_, err = encoding.WriteObject(conn, modules.AcceptContractResponse)
+		if err != nil {
+			return
+		}
 
-	// Download file contents.
-	_, err = io.CopyN(file, conn, int64(contract.FileSize))
-	if err != nil {
-		return
-	}
+		// Create file.
+		h.mu.Lock()
+		filename := h.nextFilename()
+		fullname := filepath.Join(h.hostDir, filename)
+		h.mu.Unlock()
+		file, err := os.Create(fullname)
+		if err != nil {
+			return
+		}
+		defer file.Close()
 
-	// Check that the file matches the merkle root in the contract.
-	_, err = file.Seek(0, 0)
-	if err != nil {
-		return
-	}
-	merkleRoot, err := hash.ReaderMerkleRoot(file, hash.CalculateSegments(contract.FileSize))
-	if err != nil {
-		return
-	}
-	if merkleRoot != contract.FileMerkleRoot {
-		err = errors.New("uploaded file has wrong merkle root")
-		return
-	}
+		// If there's an error upon return, delete the file that's been created.
+		defer func() {
+			if err != nil {
+				_, err = encoding.WriteObject(conn, err.Error())
+				os.Remove(fullname)
+				h.mu.Lock()
+				h.spaceRemaining -= int64(contract.FileSize)
+				h.mu.Unlock()
+			}
+		}()
 
-	// Check that the file arrived in time.
-	if h.state.Height() >= contract.Start-2 {
-		err = errors.New("file not uploaded in time, refusing to go forward with contract")
-		return
-	}
+		// Download file contents.
+		_, err = io.CopyN(file, conn, int64(contract.FileSize))
+		if err != nil {
+			return
+		}
 
-	// Put the contract in a list where the host will be performing proofs of
-	// storage.
-	h.mu.Lock()
-	h.contracts[txn.FileContractID(0)] = contractObligation{
-		filename: filename,
-	}
-	h.mu.Unlock()
-	fmt.Println("Accepted contract")
+		// Check that the file matches the merkle root in the contract.
+		_, err = file.Seek(0, 0)
+		if err != nil {
+			return
+		}
+		merkleRoot, err := hash.ReaderMerkleRoot(file, hash.CalculateSegments(contract.FileSize))
+		if err != nil {
+			return
+		}
+		if merkleRoot != contract.FileMerkleRoot {
+			err = errors.New("uploaded file has wrong merkle root")
+			return
+		}
 
-	// Submit the transaction.
-	h.state.AcceptTransaction(txn)
+		// Check that the file arrived in time.
+		if h.state.Height() >= contract.Start-2 {
+			err = errors.New("file not uploaded in time, refusing to go forward with contract")
+			return
+		}
+
+		// Put the contract in a list where the host will be performing proofs of
+		// storage.
+		h.mu.Lock()
+		h.contracts[txn.FileContractID(0)] = contractObligation{
+			filename: filename,
+		}
+		h.mu.Unlock()
+		fmt.Println("Accepted contract")
+
+		// Submit the transaction.
+		h.state.AcceptTransaction(txn)
+	*/
 
 	return
 }
