+++ modules/hostdb/entries.go
